refactor(bot): move session cluster out of utils into services
Third bounded slice of #1968. Moves sessionSnapshots.ts to
services/musicRecommendation/, and sessionStartupRestore.ts +
namedSessions.ts to services/musicManagement/ -- both of the latter
depend on sessionSnapshots, now a cross-service import mirroring the
same-directory coupling that already existed. Updates 11 import
sites plus watchdog.ts/spec (from #1981) which also depends on
sessionSnapshots.

Found and filed separately while scoping this slice, not fixed here:
#1983 (restoreSessionsOnStartup is never called -- dead code) and
#1984 (utils/music/index.ts barrel has zero external consumers).

Stacks on refactor/bot-collaborative-playlist-to-services (#1982).
