+++ packages/bot/src/functions/music/commands/session.ts
@@ -12,7 +12,7 @@ import {
     requireGuild,
     requireVoiceChannel,
 } from '../../../utils/command/commandValidations'
-import { namedSessionService } from '../../../utils/music/namedSessions'
+import { namedSessionService } from '../../../services/musicManagement/namedSessions'
 import { createQueue, queueConnect } from '../../../handlers/queueHandler'
 import { resolveGuildQueue } from '../../../utils/music/queueResolver'
