+++ packages/bot/src/services/musicRecommendation/sessionSnapshots.ts
@@ -5,7 +5,7 @@ import { randomUUID } from 'crypto'
 import { getPrismaClient, debugLog, errorLog } from '@lucky/shared/utils'
 import type { Prisma } from '@lucky/shared/utils'
 import { ENVIRONMENT_CONFIG } from '@lucky/shared/config'
-import { trackSource } from './trackFields'
+import { trackSource } from '../../utils/music/trackFields'
 
 /**
  * Extracts the structured Prisma error fields (`code` + `meta`) for logging.
